idlib x64 C++20 Warning Level 3
Builds successfully

+++ neo/idlib/math/Simd_Generic.cpp
@@ -62,7 +62,7 @@ const char * idSIMD_Generic::GetName() const {
 idSIMD_Generic::MinMax
 ============
 */
-void VPCALL idSIMD_Generic::MinMax( float &min, float &max, const float *src, const int count ) {
+void VPCALL idSIMD_Generic::MinMax( float &min, float &max, const float *src, const size_t count ) {
 	min = idMath::INFINITY; max = -idMath::INFINITY;
 #define OPER(X) if ( src[(X)] < min ) {min = src[(X)];} if ( src[(X)] > max ) {max = src[(X)];}
 	UNROLL1(OPER)
@@ -74,7 +74,7 @@ void VPCALL idSIMD_Generic::MinMax( float &min, float &max, const float *src, co
 idSIMD_Generic::MinMax
 ============
 */
-void VPCALL idSIMD_Generic::MinMax( idVec2 &min, idVec2 &max, const idVec2 *src, const int count ) {
+void VPCALL idSIMD_Generic::MinMax( idVec2 &min, idVec2 &max, const idVec2 *src, const size_t count ) {
 	min[0] = min[1] = idMath::INFINITY; max[0] = max[1] = -idMath::INFINITY;
 #define OPER(X) const idVec2 &v = src[(X)]; if ( v[0] < min[0] ) { min[0] = v[0]; } if ( v[0] > max[0] ) { max[0] = v[0]; } if ( v[1] < min[1] ) { min[1] = v[1]; } if ( v[1] > max[1] ) { max[1] = v[1]; }
 	UNROLL1(OPER)
@@ -86,7 +86,7 @@ void VPCALL idSIMD_Generic::MinMax( idVec2 &min, idVec2 &max, const idVec2 *src,
 idSIMD_Generic::MinMax
 ============
 */
-void VPCALL idSIMD_Generic::MinMax( idVec3 &min, idVec3 &max, const idVec3 *src, const int count ) {
+void VPCALL idSIMD_Generic::MinMax( idVec3 &min, idVec3 &max, const idVec3 *src, const size_t count ) {
 	min[0] = min[1] = min[2] = idMath::INFINITY; max[0] = max[1] = max[2] = -idMath::INFINITY;
 #define OPER(X) const idVec3 &v = src[(X)]; if ( v[0] < min[0] ) { min[0] = v[0]; } if ( v[0] > max[0] ) { max[0] = v[0]; } if ( v[1] < min[1] ) { min[1] = v[1]; } if ( v[1] > max[1] ) { max[1] = v[1]; } if ( v[2] < min[2] ) { min[2] = v[2]; } if ( v[2] > max[2] ) { max[2] = v[2]; }
 	UNROLL1(OPER)
@@ -98,7 +98,7 @@ void VPCALL idSIMD_Generic::MinMax( idVec3 &min, idVec3 &max, const idVec3 *src,
 idSIMD_Generic::MinMax
 ============
 */
-void VPCALL idSIMD_Generic::MinMax( idVec3 &min, idVec3 &max, const idDrawVert *src, const int count ) {
+void VPCALL idSIMD_Generic::MinMax( idVec3 &min, idVec3 &max, const idDrawVert *src, const size_t count ) {
 	min[0] = min[1] = min[2] = idMath::INFINITY; max[0] = max[1] = max[2] = -idMath::INFINITY;
 #define OPER(X) const idVec3 &v = src[(X)].xyz; if ( v[0] < min[0] ) { min[0] = v[0]; } if ( v[0] > max[0] ) { max[0] = v[0]; } if ( v[1] < min[1] ) { min[1] = v[1]; } if ( v[1] > max[1] ) { max[1] = v[1]; } if ( v[2] < min[2] ) { min[2] = v[2]; } if ( v[2] > max[2] ) { max[2] = v[2]; }
 	UNROLL1(OPER)
@@ -110,7 +110,7 @@ void VPCALL idSIMD_Generic::MinMax( idVec3 &min, idVec3 &max, const idDrawVert *
 idSIMD_Generic::MinMax
 ============
 */
-void VPCALL idSIMD_Generic::MinMax( idVec3 &min, idVec3 &max, const idDrawVert *src, const triIndex_t *indexes, const int count ) {
+void VPCALL idSIMD_Generic::MinMax( idVec3 &min, idVec3 &max, const idDrawVert *src, const triIndex_t *indexes, const size_t count ) {
 	min[0] = min[1] = min[2] = idMath::INFINITY; max[0] = max[1] = max[2] = -idMath::INFINITY;
 #define OPER(X) const idVec3 &v = src[indexes[(X)]].xyz; if ( v[0] < min[0] ) { min[0] = v[0]; } if ( v[0] > max[0] ) { max[0] = v[0]; } if ( v[1] < min[1] ) { min[1] = v[1]; } if ( v[1] > max[1] ) { max[1] = v[1]; } if ( v[2] < min[2] ) { min[2] = v[2]; } if ( v[2] > max[2] ) { max[2] = v[2]; }
 	UNROLL1(OPER)
@@ -122,7 +122,7 @@ void VPCALL idSIMD_Generic::MinMax( idVec3 &min, idVec3 &max, const idDrawVert *
 idSIMD_Generic::Memcpy
 ================
 */
-void VPCALL idSIMD_Generic::Memcpy( void *dst, const void *src, const int count ) {
+void VPCALL idSIMD_Generic::Memcpy( void *dst, const void *src, const size_t count ) {
 	memcpy( dst, src, count );
 }
 
@@ -131,7 +131,7 @@ void VPCALL idSIMD_Generic::Memcpy( void *dst, const void *src, const int count 
 idSIMD_Generic::Memset
 ================
 */
-void VPCALL idSIMD_Generic::Memset( void *dst, const int val, const int count ) {
+void VPCALL idSIMD_Generic::Memset( void *dst, const int val, const size_t count ) {
 	memset( dst, val, count );
 }
 
@@ -140,9 +140,9 @@ void VPCALL idSIMD_Generic::Memset( void *dst, const int val, const int count ) 
 idSIMD_Generic::BlendJoints
 ============
 */
-void VPCALL idSIMD_Generic::BlendJoints( idJointQuat *joints, const idJointQuat *blendJoints, const float lerp, const int *index, const int numJoints ) {
-	for ( int i = 0; i < numJoints; i++ ) {
-		const int j = index[i];
+void VPCALL idSIMD_Generic::BlendJoints( idJointQuat *joints, const idJointQuat *blendJoints, const float lerp, const size_t *index, const size_t numJoints ) {
+	for (size_t i = 0; i < numJoints; i++ ) {
+		const size_t j = index[i];
 		joints[j].q.Slerp( joints[j].q, blendJoints[j].q, lerp );
 		joints[j].t.Lerp( joints[j].t, blendJoints[j].t, lerp );
 		joints[j].w = 0.0f;
@@ -154,9 +154,9 @@ void VPCALL idSIMD_Generic::BlendJoints( idJointQuat *joints, const idJointQuat 
 idSIMD_Generic::BlendJointsFast
 ============
 */
-void VPCALL idSIMD_Generic::BlendJointsFast( idJointQuat *joints, const idJointQuat *blendJoints, const float lerp, const int *index, const int numJoints ) {
-	for ( int i = 0; i < numJoints; i++ ) {
-		const int j = index[i];
+void VPCALL idSIMD_Generic::BlendJointsFast( idJointQuat *joints, const idJointQuat *blendJoints, const float lerp, const size_t *index, const size_t numJoints ) {
+	for (size_t i = 0; i < numJoints; i++ ) {
+		const size_t j = index[i];
 		joints[j].q.Lerp( joints[j].q, blendJoints[j].q, lerp );
 		joints[j].t.Lerp( joints[j].t, blendJoints[j].t, lerp );
 		joints[j].w = 0.0f;
@@ -168,8 +168,8 @@ void VPCALL idSIMD_Generic::BlendJointsFast( idJointQuat *joints, const idJointQ
 idSIMD_Generic::ConvertJointQuatsToJointMats
 ============
 */
-void VPCALL idSIMD_Generic::ConvertJointQuatsToJointMats( idJointMat *jointMats, const idJointQuat *jointQuats, const int numJoints ) {
-	for ( int i = 0; i < numJoints; i++ ) {
+void VPCALL idSIMD_Generic::ConvertJointQuatsToJointMats( idJointMat *jointMats, const idJointQuat *jointQuats, const size_t numJoints ) {
+	for (size_t i = 0; i < numJoints; i++ ) {
 		jointMats[i].SetRotation( jointQuats[i].q.ToMat3() );
 		jointMats[i].SetTranslation( jointQuats[i].t );
 	}
@@ -180,8 +180,8 @@ void VPCALL idSIMD_Generic::ConvertJointQuatsToJointMats( idJointMat *jointMats,
 idSIMD_Generic::ConvertJointMatsToJointQuats
 ============
 */
-void VPCALL idSIMD_Generic::ConvertJointMatsToJointQuats( idJointQuat *jointQuats, const idJointMat *jointMats, const int numJoints ) {
-	for ( int i = 0; i < numJoints; i++ ) {
+void VPCALL idSIMD_Generic::ConvertJointMatsToJointQuats( idJointQuat *jointQuats, const idJointMat *jointMats, const size_t numJoints ) {
+	for (size_t i = 0; i < numJoints; i++ ) {
 		jointQuats[i] = jointMats[i].ToJointQuat();
 	}
 }
@@ -191,8 +191,8 @@ void VPCALL idSIMD_Generic::ConvertJointMatsToJointQuats( idJointQuat *jointQuat
 idSIMD_Generic::TransformJoints
 ============
 */
-void VPCALL idSIMD_Generic::TransformJoints( idJointMat *jointMats, const int *parents, const int firstJoint, const int lastJoint ) {
-	for ( int i = firstJoint; i <= lastJoint; i++ ) {
+void VPCALL idSIMD_Generic::TransformJoints( idJointMat *jointMats, const size_t *parents, const size_t firstJoint, const size_t lastJoint ) {
+	for (size_t i = firstJoint; i <= lastJoint; i++ ) {
 		assert( parents[i] < i );
 		jointMats[i] *= jointMats[parents[i]];
 	}
@@ -203,8 +203,8 @@ void VPCALL idSIMD_Generic::TransformJoints( idJointMat *jointMats, const int *p
 idSIMD_Generic::UntransformJoints
 ============
 */
-void VPCALL idSIMD_Generic::UntransformJoints( idJointMat *jointMats, const int *parents, const int firstJoint, const int lastJoint ) {
-	for ( int i = lastJoint; i >= firstJoint; i-- ) {
+void VPCALL idSIMD_Generic::UntransformJoints( idJointMat *jointMats, const size_t *parents, const size_t firstJoint, const size_t lastJoint ) {
+	for ( size_t i = lastJoint; i >= firstJoint; i-- ) {
 		assert( parents[i] < i );
 		jointMats[i] /= jointMats[parents[i]];
 	}
